:recycle: refactor: 홈 배너 emotion 스타일 컴포넌트 분리
