+++ src/widget/home/ui/index.ts
@@ -2,3 +2,4 @@ export { default as DiscoveryBanner } from './DiscoveryBanner.tsx';
 export { default as MainBanner } from './MainBanner.tsx';
 export { default as PictureWorkList } from './PictureWorkList.tsx';
 export { default as HomeFooter } from './HomeFooter.tsx';
+export * as S from './Home.styles.ts';
